Update Slurm script to set snakemake working directory
Adds a second required positional argument to determine
output/working directory. Adds an argument to singularity to
bind the workflow source into the container. This is to allow
using a python script from the workflow in the container.

I also removed the `chmod -R 774 *` logic. This seemed like a
fix for some problem related to copying the files around.